make Keyset.values' paremeter should be Int

+++ src/commonMain/kotlin/io/github/tokuhirom/kdary/internal/DoubleArrayBuilder.kt
@@ -45,7 +45,7 @@ internal class DoubleArrayBuilder(
     private fun buildDawg(keyset: Keyset): Dawg {
         val dawgBuilder = DawgBuilder()
         for (i: SizeType in 0uL until keyset.numKeys().toSizeType()) {
-            dawgBuilder.insert(keyset.keys(i.toInt()), keyset.values(i))
+            dawgBuilder.insert(keyset.keys(i.toInt()), keyset.values(i.toInt()))
             progressCallback?.invoke(i + 1uL, keyset.numKeys().toSizeType() + 1uL)
         }
         return dawgBuilder.finish()
@@ -213,12 +213,12 @@ internal class DoubleArrayBuilder(
         for (i: SizeType in begin until end) {
             val label: UByte = keyset.keys(i.toInt(), depth.toInt())
             if (label == 0.toUByte()) {
-                if (keyset.values(i) < 0) {
+                if (keyset.values(i.toInt()) < 0) {
                     throw IllegalArgumentException("failed to build double-array: negative value")
                 }
 
                 if (vaue == -1) {
-                    vaue = keyset.values(i)
+                    vaue = keyset.values(i.toInt())
                 }
                 progressCallback?.invoke(i + 1uL, (keyset.numKeys() + 1).toSizeType())
             }
