+++ src/commonTest/kotlin/io/github/tokuhirom/kdary/internal/DawgBuilderTest.kt
@@ -15,7 +15,7 @@ class DawgBuilderTest {
 
         val builder = DawgBuilder()
         for (i: SizeType in 0uL until keyset.numKeys().toSizeType()) {
-            builder.insert(keyset.keys(i.toInt()), keyset.values(i))
+            builder.insert(keyset.keys(i.toInt()), keyset.values(i.toInt()))
         }
 
         for (i: SizeType in 0uL until builder.nodes.size.toSizeType()) {
